Refactor import service utilities
